# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LoginOptionsResponse": {
        "properties": {
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_name": {
            "title": "Event Name",
            "type": "string"
          },
          "rooms": {
            "items": {
              "$ref": "#/components/schemas/LoginRoomOption"
            },
            "maxItems": 500,
            "title": "Rooms",
            "type": "array"
          }
        },
        "required": [
          "event_id",
          "event_name",
          "rooms"
        ],
        "title": "LoginOptionsResponse",
        "type": "object"
      },
...
      "RoomIdentity": {
        "properties": {
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "kind": {
            "const": "room",
            "title": "Kind",
            "type": "string"
          },
          "room_id": {
            "title": "Room Id",
            "type": "string"
          },
          "room_name": {
            "title": "Room Name",
            "type": "string"
          },
          "surface": {
            "enum": [
              "display",
              "control"
            ],
            "title": "Surface",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "room_id",
          "room_name",
          "surface",
          "event_id"
        ],
        "title": "RoomIdentity",
        "type": "object"
      },
      "RoomLoginRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "room_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Room Id",
            "type": "string"
          },
          "surface": {
            "enum": [
              "display",
              "control"
            ],
            "title": "Surface",
            "type": "string"
          }
        },
        "required": [
          "room_id",
          "password",
          "surface"
        ],
        "title": "RoomLoginRequest",
        "type": "object"
      },
...
      "Surface": {
        "enum": [
          "display",
          "control",
          "staff"
        ],
        "title": "Surface",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "control": {
        "in": "cookie",
        "name": "control_sid",
        "type": "apiKey"
      },
      "display": {
        "in": "cookie",
        "name": "display_sid",
        "type": "apiKey"
      },
      "staff": {
        "in": "cookie",
        "name": "staff_sid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Contract for room devices and staff. Prose: docs/protocol.md.",
    "title": "Proctor Suite protocol",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/logout": {
      "post": {
        "description": "Clears the cookie for one surface.",
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "in": "query",
            "name": "surface",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Surface"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unauthenticated / invalid_credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unknown_room / unknown_session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "invalid_request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "rate_limited"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/room-login": {
      "post": {
        "description": "Room name (dropdown) + event password. Sets display_sid or control_sid.",
        "operationId": "room_login_api_auth_room_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomIdentity"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unauthenticated / invalid_credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unknown_room / unknown_session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "invalid_request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "rate_limited"
          }
        },
        "summary": "Room Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/rooms": {
      "get": {
        "description": "Rooms for the login dropdown. No auth. Bounded (MAX_LOGIN_ROOMS).",
        "operationId": "login_options_api_auth_rooms_get",
        "parameters": [
          {
            "in": "query",
            "name": "practice",
            "required": false,
            "schema": {
              "default": false,
              "title": "Practice",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOptionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unauthenticated / invalid_credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unknown_room / unknown_session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "invalid_request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "rate_limited"
          }
        },
        "summary": "Login Options",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/staff-login": {
      "post": {
        "description": "Named staff account. Sets staff_sid.",
        "operationId": "staff_login_api_auth_staff_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffIdentity"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unauthenticated / invalid_credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unknown_room / unknown_session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "invalid_request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "rate_limited"
          }
        },
        "summary": "Staff Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/super-login": {
      "post": {
        "description": "Sign in with a Google ID token. Only emails on the super-admin list get a session.",
        "operationId": "super_login_api_auth_super_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuperLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuperIdentity"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "invalid_credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "not_allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "not_configured"
          }
        },
        "summary": "Super Login",
        "tags": [
          "super"
        ]
      }
    },
...
    "/api/commands": {
      "post": {
        "description": "One command, idempotent on command_id. Needs X-Proctor-Client.",
        "operationId": "post_command_api_commands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "adjust": "#/components/schemas/AdjustCommand",
                    "end": "#/components/schemas/EndCommand",
                    "pause": "#/components/schemas/PauseCommand",
                    "permit": "#/components/schemas/PermitCommand",
                    "resume": "#/components/schemas/ResumeCommand",
                    "start": "#/components/schemas/StartCommand"
                  },
                  "propertyName": "type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/PermitCommand"
                  },
                  {
                    "$ref": "#/components/schemas/StartCommand"
                  },
                  {
                    "$ref": "#/components/schemas/PauseCommand"
                  },
                  {
                    "$ref": "#/components/schemas/ResumeCommand"
                  },
                  {
                    "$ref": "#/components/schemas/EndCommand"
                  },
                  {
                    "$ref": "#/components/schemas/AdjustCommand"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unauthenticated / invalid_credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unknown_room / unknown_session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "invalid_request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "rate_limited"
          }
        },
        "summary": "Post Command",
        "tags": [
          "commands"
        ]
      }
    },
    "/api/me": {
      "get": {
        "description": "Who is logged in on this surface (cookies are HttpOnly, so the UI asks).",
        "operationId": "me_api_me_get",
        "parameters": [
          {
            "in": "query",
            "name": "surface",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Surface"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "room": "#/components/schemas/RoomIdentity",
                      "staff": "#/components/schemas/StaffIdentity"
                    },
                    "propertyName": "kind"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/RoomIdentity"
                    },
                    {
                      "$ref": "#/components/schemas/StaffIdentity"
                    }
                  ],
                  "title": "Response Me Api Me Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unauthenticated / invalid_credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "unknown_room / unknown_session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "invalid_request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "rate_limited"
          }
        },
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}